requirements.txt
flask
face_recognition
Pillow